Move S3 file system to separate package
