+++ presto-hive/src/main/java/com/facebook/presto/hive/s3/PrestoS3FileSystemStats.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive;
+package com.facebook.presto.hive.s3;
 
 import com.amazonaws.AbortedException;
 import io.airlift.stats.CounterStat;
