+++ presto-hive/src/main/java/com/facebook/presto/hive/s3/PrestoS3SseType.java
@@ -11,17 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive;
+package com.facebook.presto.hive.s3;
 
-// These are the exact names used by SignerFactory in the AWS library
-// and thus cannot be renamed or use the normal naming convention.
-@SuppressWarnings("EnumeratedConstantNamingConvention")
-public enum PrestoS3SignerType
+public enum PrestoS3SseType
 {
-    S3SignerType,
-    AWS3SignerType,
-    AWS4SignerType,
-    AWSS3V4SignerType,
-    CloudFrontSignerType,
-    QueryStringSignerType,
+    KMS,
+    S3;
 }
